Gripper action with radiant and linear interpolation for scissors
all calculations for finger position is switched from meters to radiant
to integrate it into moveit.
The position for the scissors are calculated with a linear formel out of
the state of the fingers. The finger position value is between 6 and
113, the scissor position value is between 15 and 212.

The speed for the scissors is set to maximum(255).
The speed for the fingers is set to 80.
With this combination the hand will close properly.

The name of the gripper action server is switched to "gripper_action"