fixed sending successed message
The Gripper sends the successed motion message imediatly, also when the
gripper not finneshed the motion.
The reasen is that the register gSTA was not checked if it is not 0.
If it is 0 the gripper still moves.
Also the register gIMC should be checked if it is 0x03. It shows that
the activation and mode changes are complete.