Allow different gripper modes
This merges the two branches master and basic_full_grip.

It is now possible to start the action server with the two different gripper
modes basic and wide_pinch.

It is no longer possible to set the min and max arguments for the
opening and closing radiant and effort. These values are specific for
the grasp modes and should not be changeable.