Rewrote big parts of the action server
While previously a mode had to be set at launch, now the mode is set
depending on the chosen target.
The server was changed to a FollowJointTrajectory action server.
The last point of the trajectory is compared to hard coded joint states
that correspond the target joint states of the different grasp types.
Each target joint state has a corresponding hard coded GripperOutput
that is then published.